+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Core Loot Manager v5.4.2
+## Fixes
+### Classic Era
+Alt-main linking in Classic Era should now be detected properly.
+### Tooltip
+Cataclysm tooltips should now be working again.
+### Export
+Exporting should now be working again.
+# Core Loot Manager v5.4.1
+## Fixes
+### Profiles UI
+Fixed profiles ui script errors
 # Core Loot Manager v5.4.0
 ## Changes
 ### Code unification